move computeEnrichedTerms into its own file
