+++ packages/react-tei/src/unitex/enrichDocumentWithUnitex.ts
@@ -2,48 +2,15 @@ import { kebabCasify } from "../helper/kebabCasify";
 import { mapTargetTags } from "../helper/mapTargetTags";
 import { termToRegex } from "../helper/termToRegex";
 import type { DocumentJson } from "../parser/document";
-import { getOverlappingTermsFromList } from "./getOverlappingTermsFromList";
+import { computeEnrichedTerms } from "./computeEnrichedTerms";
 import {
 	type HighlightTag,
 	highlightTermsInTextTag,
 	isTextTag,
 	type TextTag,
 } from "./highlightTermsInTextTag";
-import { mergeIdenticalTerms } from "./mergeIdenticalTerms";
-import { nestContainedTerms } from "./nestContainedTerms";
 import type { TermStatistic } from "./parseUnitexEnrichment";
-import type { NestedTerm, NormalizedTerm } from "./types";
-
-// Step 1: Normalize terms from termByGroup into flat list
-const normalizeTerms = (
-	termByGroup: Record<string, TermStatistic[]>,
-): NormalizedTerm[] => {
-	const normalized: NormalizedTerm[] = Object.entries(termByGroup).flatMap(
-		([group, terms]) =>
-			terms.map((termStat) => ({
-				term: termStat.term,
-				group,
-			})),
-	);
-
-	return normalized;
-};
-
-export const computeEnrichedTerms = (
-	termByGroup: Record<string, TermStatistic[]>,
-): NestedTerm[] => {
-	const normalizedTerms = normalizeTerms(termByGroup);
-
-	const overlappingTerms = getOverlappingTermsFromList(normalizedTerms);
-
-	const mergedNormalizedTerms = mergeIdenticalTerms(
-		normalizedTerms.concat(overlappingTerms),
-	);
-
-	const nestedTerms = nestContainedTerms(mergedNormalizedTerms);
-
-	return nestedTerms.sort((a, b) => b.term.length - a.term.length);
-};
+import type { NestedTerm } from "./types";
 
 export const termToTag = (term: NestedTerm): TextTag | HighlightTag => {
 	if (term.groups.length === 0 && !term.subTerms?.length) {
